Reseparates Schema and Type
