+++ src/runtime/particle-spec.ts
@@ -13,7 +13,8 @@ import {Modality} from './modality.js';
 import {Direction, ParticleClaimStatement, ParticleCheckStatement} from './manifest-ast-nodes.js';
 import {TypeChecker} from './recipe/type-checker.js';
 import {TypeVariableInfo} from './type-variable-info.js';
-import {Schema, InterfaceType, CollectionType, SlotType, Type, TypeLiteral} from './type.js';
+import {Schema} from './schema.js';
+import {InterfaceType, CollectionType, SlotType, Type, TypeLiteral} from './type.js';
 import {Literal} from './hot.js';
 import {Check, createCheck} from './particle-check.js';
 import {ParticleClaim, Claim, createParticleClaim} from './particle-claim.js';
