+++ src/runtime/recipe/recipe.ts
@@ -12,7 +12,8 @@ import {assert} from '../../platform/assert-web.js';
 import {digest} from '../../platform/digest-web.js';
 import {Modality} from '../modality.js';
 import {HandleConnectionSpec} from '../particle-spec.js';
-import {Schema, InterfaceType, Type} from '../type.js';
+import {Schema} from '../schema.js';
+import {InterfaceType, Type} from '../type.js';
 
 import {ConnectionConstraint, EndPoint} from './connection-constraint.js';
 import {Direction} from '../manifest-ast-nodes.js';
