+++ src/runtime/tests/data-layer-test.ts
@@ -12,7 +12,8 @@ import {assert} from '../../platform/chai-web.js';
 import {Arc} from '../arc.js';
 import {Loader} from '../../platform/loader.js';
 import {FakeSlotComposer} from '../testing/fake-slot-composer.js';
-import {Schema, EntityType} from '../type.js';
+import {Schema} from '../schema.js';
+import {EntityType} from '../type.js';
 import {ArcId} from '../id.js';
 import {collectionHandleForTest} from '../testing/handle-for-test.js';
