+++ src/runtime/tests/schema-test.ts
@@ -15,8 +15,9 @@
 import {assert} from '../../platform/chai-web.js';
 import {Manifest} from '../manifest.js';
 import {Reference} from '../reference.js';
+import {Schema} from '../schema.js';
 import {StubLoader} from '../testing/stub-loader.js';
-import {Schema, EntityType, ReferenceType} from '../type.js';
+import {EntityType, ReferenceType} from '../type.js';
 import {Entity} from '../entity.js';
 
 // Modifies the schema in-place.
