+++ src/runtime/tests/storage-proxy-test.ts
@@ -12,10 +12,11 @@
 import {assert} from '../../platform/chai-web.js';
 import {handleFor, HandleOld, Singleton, Collection} from '../handle.js';
 import {ArcId, IdGenerator} from '../id.js';
+import {Schema} from '../schema.js';
 import {StorageProxy, StorageProxyScheduler, CollectionProxy, BigCollectionProxy, SingletonProxy, NoOpStorageProxy} from '../storage-proxy.js';
 import {CrdtCollectionModel} from '../storage/crdt-collection-model.js';
 import {VolatileStorage} from '../storage/volatile-storage.js';
-import {Schema, EntityType} from '../type.js';
+import {EntityType} from '../type.js';
 import {Entity} from '../entity.js';
 import {Particle} from '../particle.js';
 import {floatingPromiseToAudit} from '../util.js';
