+++ src/tools/schema2graph.ts
@@ -7,8 +7,9 @@
  * subject to an additional IP rights grant found at
  * http://polymer.github.io/PATENTS.txt
  */
-import {Schema} from '../runtime/type.js';
+import {Schema} from '../runtime/schema.js';
 import {ParticleSpec} from '../runtime/particle-spec.js';
+import {Dictionary} from '../runtime/hot.js';
 
 export class SchemaNode {
   schema: Schema;
